Second version of ncafe post from 2016-12-23 (Poincaré conjecture)
Allow for the fact that further pairs aa^-1 or a^-1a may be removed
when sliding to get rid of one such pair, and also that the order in
which a and other arcs are passed under may change.

Require that either a or b is the arc on which p lies when realising
ab in the case for general links.